refactor : Enum 리패키징
