+++ src/test/java/webserver/request/RequestBodyParsingStrategyFactoryTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static webserver.common.MediaType.APPLICATION_FORM_URLENCODED;
+import static webserver.enums.MediaType.APPLICATION_FORM_URLENCODED;
 
 
 @DisplayName("RequestBodyParsingStrategyFactory 관련 테스트")
